+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- `:arrow_right` icon
 ### Changed
 ### Fixed