+++ CHANGELOG.md
@@ -11,7 +11,11 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 - Text: new option `wrapper_classes` to apply a class to the wrapper element
 ### Changed
 - The Input's `valid_filename` pattern now also excludes ampersands from valid filenames
-- Refactored code to use MethodBuilder and list node_adders to define available nodes in the DSL
+- Refactored code to use MethodBuilder and call `build_methods_for` to define available methods that add nodes to a control
+- A "close" button on a view-only form renders with a blue background instead of green
+- Icon includes a new `:dropdown` icon
+- DropdownButton renders with a dropdown indicator to the right of the button text
+- CallbackSection: Added methods `caption_text` and `callback_url` which act the same as `caption=` and `url=`
 ### Fixed
 
 ## [2.2.1] - 2024-03-26